Finally rename the utils file to basis_utils.h
